Assign infiniband devices under the assumption that work is equally distributed across all GPUs
Previously, when using MP=1, some gpus may decide to use the same infiniband devices, since the groups can't see the other groups. We get a better distribution of use for typical loads if we instead assume all gpus are using their infiniband devices, and thus map infiniband devices to gpus in a deterministic manner, regardless of which gpus are part of the current group.