processgroup: mutex -> threadUnsafe, fix resetStepValue
ThreadUnsafe will throw an exception if concurrent use is detected. Taking a mutex is a bit silly since it does not make us thread-safe (due to stepValue). Nonetheless, ThreadUnsafe is unlikely to trigger since the calls are non-blocking, but it's something.
freeMemory also does not need to take a mutex, since ipcMapper is internally synchronized, and sync is always carefully called *before* any ipcMapper ops. This fixes a resetStepValue deadlock.