cuMemImportFromShareableHandle: invalid device ordinal
cuMemImportFromShareableHandle straight-up does not work for IPC (possibly with the cute exception of between direct parent-child). There's no indication, in the docs or elsewhere, that it shouldn't work - if anything, it's documented as if it really should work. Cuda support forums have plenty of threads of other people having the same issue, with no official response. Typical.