Wheels, docker build script, improved build
We can now build wheels in a docker container with docker-build-wheels.sh.

We no longer call into pytorch' cmake files, which means that we don't require CUDA to be enabled as a language, and we don't require CUDA to be working in order to build - we only require the cuda toolkit to be installed.
Removed some dependencies like libnuma